Added min value support.

+++ src/Arbor.ModelBinding.Primitives/LongValueTypeAttribute.cs
@@ -5,7 +5,8 @@ namespace Arbor.ModelBinding.Primitives
     [AttributeUsage(AttributeTargets.Class)]
     public sealed class LongValueTypeAttribute : Attribute
     {
-        public LongValueTypeAttribute(int minValue = 1) => MinValue = minValue;
-        public int MinValue { get; }
+        public LongValueTypeAttribute(long minValue = 1) => MinValue = minValue;
+
+        public long MinValue { get; }
     }
 }
